+++ flow-typed/npm/@vanilla-extract/private_vx.x.x.js
@@ -1,0 +1,5 @@
+declare module "@vanilla-extract/private" {
+    declare module.exports: {
+        CSSVarFunction: string,
+    };
+}