+++ src/__acceptance_tests__/__ssr_pages__/popover.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {Popover} from '../..';
+import {Popover} from '../../..';
 
 const PopoverTest: React.FC = () => (
     <Popover